feat($compile/ngBind): only add binding info if debugInfo is enabled
The compiler and ngBind directives add binding information (`ng-binding`
CSS class and `$binding` data object) to elements when they are bound to
the scope. This is only to aid testing and debugging for tools such as
Protractor and Batarang. In production this is unneccesary and add a
performance penalty.

We were storing the whole `interpolationFn` in the `$binding` data on
elements but this function was bringing a lot of closure variables with it
and so was consuming unwanted amounts of memory.

Now we are only storing the parsed interpolation expressions from the
binding (i.e. the values of `interpolationFn.expressions`).

This change disables adding this binding information by default. You can
enable it by calling `$compilerProvider.enableDebugInfo()` in a module
`config()` block.

BREAKING CHANGE:

The `ng-binding` CSS class and the `$binding` data object are no longer
attached to DOM elements that have a binding to the scope, by default.
If you relied on this feature then you can re-enable it in one of your
application modules:

```
someModule.config(['$compileProvider', function($compileProvider) {
  $compileProvider.enableDebugInfo();
}]);
```

Also bindings are always arrays now and they do not include the curly
braces `{{ ... }}` in the binding expression that is stored.